seccomp: Split out seccomp_types.h
mainline inclusion
from mainline-v6.8-rc1
category: performance

More pruning of sched.h dependencies.

Conflicts:
	include/linux/sched.h
(cherry picked from commit a6e1420ce4fc91da56c0a2444c4482245e7617d4)
